Refactor project
- Refactor project to separate app.py
into smaller files
- Add helpers folder
- Add blueprints folder
