+++ app/blueprints/__init__.py
@@ -1,0 +1,5 @@
+from .auth_routes import AuthRoutes
+from .spotify_routes import SpotifyRoutes
+
+
+__all__ = ["AuthRoutes", "SpotifyRoutes"]
